Medico.AI/requirements.txt
sentence-transformers==2.2.2
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain==0.2.15
langchain-google-genai==0.1.6
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
-e .
